More move / cleanup code.
